requirements.txt
Flask==2.2.3
Flask-CORS==3.0.10
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.4
Werkzeug==2.2.3
pymongo==4.4.1
python-dotenv==1.0.0
jwt==1.3.1
tensorflow==2.13.0
opencv-python==4.10.0.82
scikit-learn==1.3.0
numpy==1.23.5
cloudinary==1.34.0
requests==2.31.0
matplotlib==3.8.4
numpy==1.26.4
bson==0.5.10
